New CLI API: minor improvements. Refactoring.
* Added some more messages that adds verbose and information about API
  requests in following cases:
  1) Trying to remove the project-token with invalid id:
  2) Trying to add/remove non-existent user to/from collaborators;
  3) Trying to add/remove user to/from collaborators that has been
     already added/deleted or has never been added/deleted.
* User API improvements;
* Refactoring;
* Stylystic fixes.

+++ src/api/handlers/projects/tokens.py
@@ -1,6 +1,7 @@
-from flask import request, g
+from flask import request, g, abort
 from flask_restplus import Resource, fields
 
+from pyinfrabox.utils import validate_uuid4
 from pyinfraboxutils.ibflask import auth_required, OK
 from pyinfraboxutils.ibrestplus import api
 from pyinfraboxutils.token import encode_project_token
@@ -49,10 +50,13 @@ class Token(Resource):
 
     @auth_required(['user'])
     def delete(self, project_id, token_id):
-        g.db.execute('''
-            DELETE FROM auth_token
-            WHERE project_id = %s and id = %s
-        ''', [project_id, token_id])
+        if not validate_uuid4(token_id):
+            abort(400, "Invalid projet-token uuid")
+
+        g.db.execute("""
+                     DELETE FROM auth_token
+                     WHERE project_id = %s and id = %s
+        """, [project_id, token_id])
         g.db.commit()
 
         return OK('Successfully deleted token')
